Release v1.1.0 (#4)
* stash - moving to laptop

* Updated run to work even if the user uses full path, updated hocwriter to remove unused args, updated runner to suppress pymunk welcome message

* changing function input

* added contour smoothing

* Update README.md

* stash

* typo from recent merge

* stash - saving the AP times at either end works

* tested (backwards compatibility when "end_ap_times" is not defined)
tested error checking for min/max end locations self.throw(111)
tested NEURON with and without ap ends checking

* removed SL code

* added exception

* exception for no deform ratio

* added exceptions to prevent people (1) from selecting AP record locations not on the fiber and (2) from setting the terminal nodes to be recorded which are also grounded (that would be confusing)

* prep resolve merge conflicts

* conflict resolution

* Controls to save AP times at the ends of the fiber

* Update sample.py

* Update exceptions.json

* remove all star imports

* fixing deformation

* fixed deformation logic

* removing changes from other branch

* reverting change to error message

* cleaning up im preprocess

* added install dependencies

* added to sample template

* added exception

* added throw call

* minor, fixes breaking error: introduced during refactor for diameter distributions

* added fiber length error message

* fixed override_length not properly used by clip

* added error message to both fiber build blocks

* added error message to both build calls for fiberset

* fixed error to account for fiber min not being 0

* Guess I just learned how arrows work

* Fiber message

* update exception

* im dumb

* removed extraneous files

* Deform fix

* Image preprocessing

* m

* removes unused yml for mkdocs

* Specifies location of erroroneous parameter for de-islanding in Sample

* rmstars

* cleanup code

* added environ check

* removes unused parameter z_nerve from the CorTec cuffs

* fixes breaking error for old samples that do not have the "deform_ratio" parameter defined

* sets diams to [np.nan, ... ] if not diam_distribution, in which case the diameters are not saved with the fiber

* updated templates

* added troubleshooting guide

* minor formatting

* added ensuredirs to installation

* Clarification to nerve interpolation

* Install fix

* updated docs

* changes scope of 'diameter' variable def in logic

* Changed the peri fit equation (minor bug only in scope of deforming the fascicles ... the CI value in COMSOL is derived from the fascicle area)

Now uses the actual area to get equivalent-circle-radius (now matches Pelot et al. pub), not mean of a and b from best-fit ellipse

* fixed case where inners count >1 for a single outer would cause numpy error

* added dims check

* Exceptionable bugfix

* replaced O(n^4) naive algorithm with O(n) sampling algorithm for minimum bounding circle

* fixed rotation

* error for slide not centered

* updated error case to have proper ordering for short circuiting

* updating code number

* fixed scout sim

* fixed error messages to properly function without needing to be sequential

* fixed

* fixed

* fixed bug

* updated docs and added warning

* allow slide to be a few microns off origin due do changes in nerve trace from smoothing or deforming

* removing error for now because sometimes slide gets very far off center

* src/core/sample.py

* Revert "Merge branch 'rotation_fix' into 'master'"

This reverts merge request !113

* Orientation update

* Added logic to ensure fascicle separation

* interp curve set to closed

* added error messages for orientation tif

* Added plot folder feature

* Move center

* Rotation mod

* corrected warning message

* Update S2-Installation.md

* minor fix: interp is actually valid down to 1 um

* added requirements.txt and updated windows install

* Revert "Merge branch 'increase_mrg_interp_bounds' into 'master'"

This reverts merge request !126

* updated shell script

* updated shell script

* remove newline

* Update welcome message

* adds controls to Sim for whether the program saves the NEURON runtime

* updated query

* removed print statement from get

* fix bug with orientation point message

* Update README.md

* Basis message

* Cuff spec

* fixed windows installation

* targeted pip

* init powershell

* fixed

* fixedfr

* fixed sh

* small change

* fixed

* removes unused file

* Revert "Merge branch 'requirements' of https://gitlab.oit.duke.edu/wmglab/ascent"

This reverts commit a2aae9035032a9e1c4a0551b34fb8cb98704c203, reversing
changes made to 9010b32ec70fa5644ea708c76aa163a106dc8e5f.

* Update submit.py message

* Update installation to use requirements.txt

* Model in use fix

* cleanup_linting_warnings

* env_setup fixed

* Ellipses descriptions consistent between MockSample and Sample (leaving Sample alone, changing MockSample).

* Revert "Merge branch 'requirements' into 'master'"

This reverts merge request !127

* updated install

* added reqs file

* Update simulation.py

* Stack trace

* updated installation docs

* removed S titles

* Remove map path

* Update simulation.py

* Update barchart_sample_comparison.py

* Query mods

* generalize

* Arg parser

* looping plot

* Ap ends times docs

* moved utils

* new loc

* rm sphinx

* Adds file and instructions for import_n_sims

* Update install.sh

* Update Helper-scripts.md

* Update parse_args.py

* Release v1.1.0

* Update requirements.txt

* added error message

* Update S2-Installation.md

* Update Troubleshooting-Guide.md

* Release v1.1.0

* fixed orientation for naive modes

* rm changes

+++ src/core/simulation.py
@@ -169,7 +169,12 @@ class Simulation(Exceptionable, Configurable, Saveable):
                 plot: bool = self.search(Config.SIM, 'waveform', 'plot')
 
             if plot:
-                waveform.plot(final=True)
+                if self.search(Config.SIM, 'plot_folder',optional = True) == True: 
+                    path = sim_directory+'/plots/waveforms/{}.png'.format(i)
+                    if not os.path.exists(sim_directory+'/plots/waveforms'):
+                        os.makedirs(sim_directory+'/plots/waveforms')
+                else: path = None
+                waveform.plot(final=True,path=path)
 
             self.waveforms.append(waveform)
 
@@ -229,7 +234,7 @@ class Simulation(Exceptionable, Configurable, Saveable):
         key_file_dir = os.path.join(sim_directory, "potentials")
         key_filepath = os.path.join(sim_directory, "potentials", "key.dat")
         if not os.path.exists(key_file_dir):
-            os.mkdir(key_file_dir)
+            os.makedirs(key_file_dir)
 
         with open(key_filepath, 'w') as f:
             for row in output:
@@ -259,12 +264,9 @@ class Simulation(Exceptionable, Configurable, Saveable):
             diams = np.loadtxt(os.path.join(my_potentials_directory, my_file))
             for fiber_ind in range(len(diams)):
                 diam = diams[fiber_ind]
-
-                # NOTE: if SL interp, writes files as inner0_fiber<q>.dat
-                if not my_xy_mode == FiberXYMode.SL_PSEUDO_INTERP:
-                    inner, fiber = self.indices_fib_to_n(my_p, fiber_ind)
-                else:
-                    inner, fiber = 0, fiber_ind
+                
+                inner, fiber = self.indices_fib_to_n(my_p, fiber_ind)
+                
                 inner_fiber_diam_key.append((inner, fiber, diam))
 
             inner_fiber_diam_key_filename = os.path.join(nsim_inputs_directory, 'inner_fiber_diam_key.obj')
@@ -334,7 +336,7 @@ class Simulation(Exceptionable, Configurable, Saveable):
             hocwriter \
                 .add(SetupMode.OLD, Config.MODEL, self.configs[Config.MODEL.value]) \
                 .add(SetupMode.OLD, Config.SIM, sim_copy) \
-                .build_hoc(n_inners, n_fiber_coords[fiberset_ind], n_tsteps)
+                .build_hoc(n_tsteps)
 
             # copy in potentials data into neuron simulation data/inputs folder
             # the potentials files are matched to their inner and fiber index, and saved in destination folder with
@@ -365,10 +367,8 @@ class Simulation(Exceptionable, Configurable, Saveable):
                             # NOTE: if SL interp, writes files as inner0_fiber<q>.dat
                             l: int
                             k: int
-                            if not xy_mode == FiberXYMode.SL_PSEUDO_INTERP:
-                                l, k = self.indices_fib_to_n(p, q)
-                            else:
-                                l, k = 0, q
+                            
+                            l, k = self.indices_fib_to_n(p, q)
 
                             is_member = np.in1d(l, inner_list)
                             if not is_member:
@@ -477,10 +477,8 @@ class Simulation(Exceptionable, Configurable, Saveable):
                                             # NOTE: if SL interp, writes files as inner0_fiber<q>.dat
                                             l: int
                                             k: int
-                                            if not xy_mode == FiberXYMode.SL_PSEUDO_INTERP:
-                                                l, k = self.indices_fib_to_n(p, q)
-                                            else:
-                                                l, k = 0, q
+                                            
+                                            l, k = self.indices_fib_to_n(p, q)
 
                                             ss_filename = 'inner{}_fiber{}.dat'.format(l, k)
 
@@ -582,7 +580,7 @@ class Simulation(Exceptionable, Configurable, Saveable):
 
         # make NSIM_EXPORT_PATH (defined in Env.json) directory if it does not yet exist
         if not os.path.exists(target):
-            os.mkdir(target)
+            os.makedirs(target)
 
         # neuron files
         du.copy_tree(os.path.join(os.environ[Env.PROJECT_PATH.value], 'src', 'neuron'), target)
@@ -613,7 +611,8 @@ class Simulation(Exceptionable, Configurable, Saveable):
         for dirname in [f for f in os.listdir(source) if os.path.isdir(os.path.join(source, f))]:
             this_sample, this_model, this_sim, product_index = tuple(dirname.split('_'))
             if sample == int(this_sample) and model == int(this_model) and sim == int(this_sim):
-                shutil.rmtree(os.path.join(sim_dir, product_index))
+                if os.path.isdir(os.path.join(sim_dir, product_index)):
+                    shutil.rmtree(os.path.join(sim_dir, product_index))
                 shutil.copytree(os.path.join(source, dirname), os.path.join(sim_dir, product_index))
 
     def potentials_exist(self, sim_dir: str) -> bool:
